Refactor request/response types for node REST client
