+++ clients/node/src/index.ts
@@ -3,5 +3,6 @@ export * from './constants'
 export * from './errorTypes'
 export * from './paginationTypes'
 export * from './parseDates'
+export * from './requestTypes'
 export * from './responseTypes'
 export * from './sanitiseError'
